backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "An end-to-end research paper management system for students"
authors = [
    { name = "Harshit Sharma", email = "harshit158@gmail.com" }
]
dependencies = [
    "pymupdf>=1.25.1",
    "ipykernel>=6.29.5",
    "instructor[google-generativeai]>=1.7.2",
    "openai>=1.59.7",
    "pydantic>=2.10.5",
    "python-dotenv>=1.0.1",
    "nltk>=3.9.1",
    "google-generativeai>=0.8.3",
    "rich>=13.9.4",
    "cohere>=5.13.8",
    "streamlit>=1.41.1",
    "langchain-text-splitters>=0.3.5",
    "pysbd>=0.3.4",
    "semantic-text-splitter>=0.22.0",
    "llmsherpa>=0.1.4",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "arxiv>=2.1.3",
    "gliner>=0.2.16",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.rye.scripts]
extract_entities = "python scripts/extract_datasets.py"
dashboard = "streamlit run dashboard/app.py"